+++ inc/UnusedMemory.h
@@ -13,10 +13,10 @@ namespace EightBit {
 	// is being read.
 	class UnusedMemory final : public Memory {
 	public:
-		UnusedMemory(size_t size, uint8_t value) noexcept;
+		UnusedMemory(uint16_t size, uint8_t value) noexcept;
 		~UnusedMemory() = default;
 
-		[[nodiscard]] size_t size() const noexcept final;
+		[[nodiscard]] uint16_t size() const noexcept final;
 		[[nodiscard]] uint8_t peek(uint16_t address) const noexcept final;
 
 		int load(std::ifstream& file, int writeOffset = 0, int readOffset = 0, int limit = -1) final;
@@ -27,7 +27,7 @@ namespace EightBit {
 		void poke(uint16_t address, uint8_t value) noexcept final;
 
 	private:
-		size_t m_size;
+		uint16_t m_size;
 		uint8_t m_value;
 	};
 }
